trilha1/curso7/requirements.txt
ipykernel
openai>=1.0.0
python-dotenv
yfinance
pandas
numpy
matplotlib
openmeteo-requests
requests-cache 
retry-requests 